AI_Employee/requirements.txt
anthropic
watchdog
mcp
requests
schedule
python-dotenv
